requirements.txt
altair==5.0.1
appdirs==1.4.4
art==6.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
decorator==5.1.1
frozendict==2.3.8
gitdb==4.0.10
GitPython==3.1.32
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multitasking==0.0.11
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.11
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.24.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
webencodings==0.5.1
yfinance==0.2.24
zipp==3.16.2
